Rename to click7, not _click7
