+++ typer/__init__.py
@@ -2,28 +2,28 @@
 
 __version__ = "0.3.2"
 
-from ._click7.exceptions import Abort as Abort
-from ._click7.exceptions import BadParameter as BadParameter
-from ._click7.exceptions import Exit as Exit
-from ._click7.termui import clear as clear
-from ._click7.termui import confirm as confirm
-from ._click7.termui import echo_via_pager as echo_via_pager
-from ._click7.termui import edit as edit
-from ._click7.termui import get_terminal_size as get_terminal_size
-from ._click7.termui import getchar as getchar
-from ._click7.termui import launch as launch
-from ._click7.termui import pause as pause
-from ._click7.termui import progressbar as progressbar
-from ._click7.termui import prompt as prompt
-from ._click7.termui import secho as secho
-from ._click7.termui import style as style
-from ._click7.termui import unstyle as unstyle
-from ._click7.utils import echo as echo
-from ._click7.utils import format_filename as format_filename
-from ._click7.utils import get_app_dir as get_app_dir
-from ._click7.utils import get_binary_stream as get_binary_stream
-from ._click7.utils import get_text_stream as get_text_stream
-from ._click7.utils import open_file as open_file
+from .click7.exceptions import Abort as Abort
+from .click7.exceptions import BadParameter as BadParameter
+from .click7.exceptions import Exit as Exit
+from .click7.termui import clear as clear
+from .click7.termui import confirm as confirm
+from .click7.termui import echo_via_pager as echo_via_pager
+from .click7.termui import edit as edit
+from .click7.termui import get_terminal_size as get_terminal_size
+from .click7.termui import getchar as getchar
+from .click7.termui import launch as launch
+from .click7.termui import pause as pause
+from .click7.termui import progressbar as progressbar
+from .click7.termui import prompt as prompt
+from .click7.termui import secho as secho
+from .click7.termui import style as style
+from .click7.termui import unstyle as unstyle
+from .click7.utils import echo as echo
+from .click7.utils import format_filename as format_filename
+from .click7.utils import get_app_dir as get_app_dir
+from .click7.utils import get_binary_stream as get_binary_stream
+from .click7.utils import get_text_stream as get_text_stream
+from .click7.utils import open_file as open_file
 
 from . import colors as colors
 from .main import Typer as Typer
